+++ src/main/java/ga/banga/calculatriceversion/exception/GlobalExceptionHandler.java
@@ -47,4 +47,14 @@ public class GlobalExceptionHandler {
 
         return new ResponseEntity<>(body, HttpStatus.INTERNAL_SERVER_ERROR);
     }
+
+    @ExceptionHandler(OperationImpossibleException.class)
+    public ResponseEntity<Map<String, Object>> handleOperationImpossibleException(OperationImpossibleException e) {
+        Map<String, Object> body = new HashMap<>();
+        body.put("timestamp", LocalDateTime.now());
+        body.put("message", e.getMessage());
+        body.put("status", HttpStatus.BAD_REQUEST.value());
+
+        return new ResponseEntity<>(body, HttpStatus.BAD_REQUEST);
+    }
 }