+++ src/main/java/ga/banga/calculatriceversion/exception/OperationImpossibleException.java
@@ -1,0 +1,10 @@
+package ga.banga.calculatriceversion.exception;
+
+public class OperationImpossibleException extends RuntimeException {
+
+    public OperationImpossibleException(String message) {
+        super(message);
+    }
+
+
+}